rjson: inject yielding into rjson parsing
Parsing large JSON documents can be a source of reactor stalls,
e.g. on a local installation, inserting an element with PutItem
can cause a 260ms stall if the element is a JSON object with
one million string fields. Unfortunately, rapidjson library
does not allow to easily override parsing routines, so this commit
injects the yield check by putting it inside a rapidjson macro.
We already override one macro (RAPIDJSON_ASSERT) for our own purposes,
since it seems to be the way rapidjson authors allow changing
their library internals.
It would be more proper to instead create a subclass of a parser,
but in rapidjson's current state it results in too much copy-pasta,
since many needed methods are private instead of protected.
In order for yielding to work, parsing must be run inside
seastar::thread, otherwise the yield check will be ignored.