rjson: add yieldable writer for printing JSON objects
rapidjson's default writer may be a cause of reactor stalls
for large JSON objects (e.g. an object with 1 million fields).
To mitigate it, a yield check is injected by providing
a thin wrapper subclass for the writer.
Note that in order for yields to work, the writer needs to be run
in seastar::thread context, otherwise it does nothing.