cql3: Parse cf name in drop_index_satement::validate
The patch 759752947b explains why the .column_family method
of this statament implementation must be tuned to calculate
the column_family in some cases. However, to do this the global
storage_proxy is needed.

The proposal is to calculate the column_family in .validate
method, like it's done e.g. for function_statement-s, which
has storage_proxy reference at hands.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>