cdc: use `extract_changes` in `transform`
This is a refactor of the `transform` function.
It doesn't inspect the mutation manually now. Instead, it calls
`extract_changes` which returns an easily accesible (structured) set of
changes, split by timestamp and ttl.

It then adds a CDC log table row for each of those changes.

As a byproduct:
1. the CDC log now distinguishes between inserts, updates, and row deletes.
2. static row changes appear even if no clustering rows were affected by
   the base mutation (the static row change is always a separate
   change).

The commit also fixes pre_image_select to query the preimage even if the
mutation contains range or partition deletions.